src/test/java/ma/zs/stocky/unit/service/impl/admin/crm/PaymentTypeAdminServiceImplTest.java
package ma.zs.stocky.unit.service.impl.admin.crm;

import ma.zs.stocky.bean.core.crm.PaymentType;
import ma.zs.stocky.dao.facade.core.crm.PaymentTypeDao;
import ma.zs.stocky.service.impl.admin.crm.PaymentTypeAdminServiceImpl;

import java.util.*;
import java.math.BigDecimal;
import java.time.LocalDateTime;
import java.util.stream.Collectors;
import java.util.stream.IntStream;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import java.time.LocalDateTime;



import org.springframework.boot.test.context.SpringBootTest;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;


@SpringBootTest
class PaymentTypeAdminServiceImplTest {

    @Mock
    private PaymentTypeDao repository;
    private AutoCloseable autoCloseable;
    private PaymentTypeAdminServiceImpl underTest;



    @BeforeEach
    void setUp() {
        autoCloseable = MockitoAnnotations.openMocks(this);
        underTest = new PaymentTypeAdminServiceImpl(repository);
    }

    @AfterEach
    void tearDown() throws Exception {
        autoCloseable.close();
    }

    @Test
    void canGetAllPaymentType() {
         //when
        underTest.findAll();
        verify(repository).findAll();
    }

    @Test
    void itShouldSavePaymentType() {
        // Given
        PaymentType toSave = constructSample(1);
        when(repository.save(toSave)).thenReturn(toSave);

        // When
        underTest.create(toSave);

        // Then
        verify(repository).save(toSave);
    }

    @Test
    void itShouldDeletePaymentType() {
        // Given
        Long idToDelete = 1L;
        when(repository.existsById(idToDelete)).thenReturn(true);

        // When
        underTest.deleteById(idToDelete);

        // Then
        verify(repository).deleteById(idToDelete);
    }
    @Test
    void itShouldGetPaymentTypeById() {
        // Given
        Long idToRetrieve = 1L; // Example PaymentType ID to retrieve
        PaymentType expected = new PaymentType(); // You need to replace PaymentType with your actual class
        expected.setId(idToRetrieve);
        when(repository.findById(idToRetrieve)).thenReturn(java.util.Optional.of(expected));

        // When
        PaymentType result = underTest.findById(idToRetrieve);

        // Then
        assertEquals(expected, result);
    }
	
	private PaymentType constructSample(int i) {
		PaymentType given = new PaymentType();
        given.setReference("reference-"+i);
        given.setCode("code-"+i);
        return given;
    }

}
